FEATURE-ADD
*Objetos de sala
(Adicionamos mais um membro da nossa gameObjectTree, o "BaseRoomGameObject", que é uma classe filha de "BaseGameObject", mas que é gerenciada pelo manager de gameobjects e pela "PlayableRoom" ao mesmo tempo)

*Marcação de objetos
(Como estamos usando um sistema de controle de objetos genérico, o sistema de "PlayableRoom" deve apenas servir como um conteiner genérico de dados)

*Limpeza de salas e atualização pra uma nova
(Com o sistema salas criado, implementamos um sistema para limpar elas e seus dados antigos não persistentes)

OBS: QUANDO ADICIONAMOS UM NOVO "BaseRoomGameObject", OU ALGUMA DE SUAS CLASSES FILHAS, ELAS AUTOMATICAMENTE SE INTRODUZEM NOS GERENCIADORES PROPRIOS, PARA O GERENCIADOR DE OBJETOS CHAMAMOS A PIPELINE QUE IRÁ PREPARAR PRA ADIÇÃO, E ADICIONAMOS DE MODO DIRETO NA LISTA DE OBJETOS DE SALA, POIS A VERDADE É QUE NA SALA ALÍ ESTAMOS MARCANDO QUEM É DE LÁ OU NÃO.

OBS_2: AS SALAS NÃO DITAM COMO OS OBJETOS SERÃO ATUALIZADO NEM NADA DO TIPO, ELAS SÃO APENAS CONTEINERS DE DADOS, PORTANTO SE QUISERMOS REMOVER UM OBJETO DE UMA SALA E QUISERMOS ELE PARE DE SER RENDERIZADO E ATUALIZADO, PRECISAMOS REMOVER ELE USANDO A PIPELINE NATURAL DOS "BaseGameObject", POIS A SOBRECRITA DE MÉTODOS A PARTIR DE "BaseRoomGameObject", INTRODUZ UMA SEQUENCIA DE DESTRUIÇÃO QUE IRÁ REMOVER O OBJETO DA LISTA DE OBJETOS DA SALA AUTOMATICAMENTE, O QUE É IMPORTANTE POIS ESSAS SALAS IRÃO MANTER DADOS DE OBJETOS DINAMICAMENTE E DEVERÁ SER MUTÁVEL. MAS SE QUISERMOS REMOVER UM OBJETO DE UMA SALA, E FAZER COM QUE ESSE OBJETO SEJA REMOVIDO DO MANAGER, TENDO APENAS ACESSO AO ROOM... AS COISAS PODEM FICAR COMPLICADAS, MAS É SÓ TERMOS CONTROLE DE QUEM PODE OU NÃO DITAR ESSAS INFORMAÇÕES QUE AS COISAS FICAM MAIS SIMPLES

OBS_3: AS SALAS JOGÁVEIS SÃO APENAS CONTEINERS DE DADOS, ELES NÃO DITAM COMO O MANAGER DE OBJETOS DEVERÁ LIDAR COM AS COISAS, NA VERDADE O QUE ACONTECE É QUE QUANDO CARREGAMOS UMA SALA, DIZEMOS PRO SISTEMA PEGAR OS DADOS DA SALA E APLICAR NO MANAGER, MAS A MAIORIA DOS OBJETOS JÁ SE AUTOAPLICAM, ENTÃO ACABA QUE O SISTEMA PASSA A LIDAR CONSIGO MESMO DE UM MODO ISOLADO, O QUE PRECISAMOS TER EM MENTE, É QUE ISSO PASSA A TORNAR A "PlayableRoom", ALGO MUITO PROXIMO DE UMA TAG COM PARAMETROS... UMA SIMPLIFICAÇÃO MEIO EXAGERADA DE FATO, MAS É O QUE MAIS FAZ SENTIDO PARA EXPLICAR SUA FUNCIONALIDADE. AGORA SE QUISERMOS CARREGAR DADOS QUE IRÃO VIR ESPECÍFICAMENTE DE UMA SALA, PRECISARIAMOS USAR ALGO QUE CHAMO DE BLUEPRINT, ESSA QUE IRIA CONTER OS DADOS QUE IRIAM PRECISAR SER CARREGADOS POR SALA, QUER SEJA O MAPING DE TILES, AS TILES, OS OBJETOS, EVENTOS E COISAS DO TIPO, QUE PRECISARÃO SER CARREGADAS NO MANAGER DE OBJETOS E REGISTRADAS COMO FAZENDO PARTE DA SALA

+++ core/src/main/java/official/sketchBook/engine/world_gen/model/PlayableRoom.java
@@ -2,7 +2,9 @@ package official.sketchBook.engine.world_gen.model;
 
 import com.badlogic.gdx.physics.box2d.Body;
 import com.badlogic.gdx.physics.box2d.World;
+import official.sketchBook.engine.gameObject_related.BaseRoomGameObject;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -33,11 +35,19 @@ public class PlayableRoom {
         roomWidthPx,
         roomHeightPx;
 
-    /// Referência ao mundo físico do box2d
+    /// Referência ao mundo físico do box2d, apenas uma referencia para determinar
+    /// se a sala tem acesso a esse mundo ou não, ela é quem determinará o fluxo da geração das tiles,
+    /// se elas irão ter um corpo físico ou não,
+    /// porém isso não dita a existencia de um mundo físico para o manager de objetos nem aos seus objetos
     private final World physicsWorld;
 
-    /// Corpos nativos da sala, não gerenciados externamente
+    /// Corpos nativos da sala, são gerenciados pelo manager de sala,
+    /// são na verdade os corpos das tiles e outros elementos nativos exclusivos da room
     public List<Body> nativeBodies;
+
+    /// Lista de referência a objetos presentes na sala, são apenas referencias,
+    /// e não são gerenciados aqui dentro
+    public List<BaseRoomGameObject> roomGameObjectList;
 
     /// Flag que dita se podemos ou acessar o world
     private final boolean physicsWorldAccessible;
@@ -50,14 +60,23 @@ public class PlayableRoom {
         float roomY,
         World physicsWorld
     ) {
-        this.roomId = id;                               //Id da sala
-        this.roomXPos = roomX;                          //Posição que iremos começar a gerar a sala no eixo x
-        this.roomYPos = roomY;                          //Posição que iremos começar a gerar a sala no eixo y
+        this.roomId = id;                                   //Id da sala
+        this.roomXPos = roomX;                              //Posição que iremos começar a gerar a sala no eixo x
+        this.roomYPos = roomY;                              //Posição que iremos começar a gerar a sala no eixo y
 
-        this.physicsWorld = physicsWorld;               //Inicializamos com o world
+        this.physicsWorld = physicsWorld;                   //Inicializamos com o world
         this.physicsWorldAccessible = physicsWorld != null; //Validamos se podemos usar o world
 
         this.tileModelIdMap = new HashMap<>();               //Inicializamos o hashMap
+        roomGameObjectList = new ArrayList<>();
+    }
+
+    public void addNewRoomGameObject(BaseRoomGameObject roomObject) {
+        this.roomGameObjectList.add(roomObject);
+    }
+
+    public void removeRoomObject(BaseRoomGameObject roomObjectToRemove) {
+        this.roomGameObjectList.remove(roomObjectToRemove);
     }
 
     public void dispose() {
@@ -70,7 +89,7 @@ public class PlayableRoom {
     }
 
     private void disposeNativeBodies() {
-        if (!physicsWorldAccessible) return;
+        if (!physicsWorldAccessible || nativeBodies == null) return;
         for (int i = nativeBodies.size() - 1; i >= 0; i--) {
             physicsWorld.destroyBody(nativeBodies.get(i));
         }
@@ -79,6 +98,7 @@ public class PlayableRoom {
     /// Limpa todas as listas
     private void disposeLists() {
         tileModelIdMap.clear();
+        roomGameObjectList.clear();
 
         if (nativeBodies != null) {
             nativeBodies.clear();
@@ -92,5 +112,17 @@ public class PlayableRoom {
 
     public World getPhysicsWorld() {
         return physicsWorld;
+    }
+
+    public float getRoomXPos() {
+        return roomXPos;
+    }
+
+    public float getRoomYPos() {
+        return roomYPos;
+    }
+
+    public int getRoomId() {
+        return roomId;
     }
 }
